fix: In servers_specs.yml, use block style rather than mixed style
Without this change, servers_specs.yml uses an odd combination of block
and flow style, resulting in very long lines that might wrap around in
the lab terminal.

It's much cleaner to use block style here, so do just that.

Also, vol_size is meant to be a numeric value, so drop the quotes
around it.