fix: Drop the reference to the ini inventory plugin
The whole point of this tutorial is to use the OpenStack dynamic
inventory plugin. We have no use for the ini plugin, and we don't need
it either: the only host besides the OpenStack servers is localhost,
and localhost is always implicitly included in the inventory.

Drop the reference to the ini plugin, and include the openstack.yml
plugin directly via an inventory option in ansible.cfg.