fix: Add the only_clouds setting to the openstack.yml configuration
If this repository is used only in the context of the tutorial,
omitting only_clouds is fine.

However if the learner chooses to use the clone the repo locally, and
wants to run the inventory plugin against their own cloud environment,
then the inventory driver will collect data from **all** sources
defined in the learner's clouds.yaml file.

By setting only_clouds to the empty list, we make sure that the
inventory plugin honours just the information that is set via
environment variables.