refactor: Properly use the inventory driver for distro discovery
There is no real need to specify the default user explicitly. Rather,
we can use the inventory driver's rich metadata to tell us what
distribution a server was built from, without ever connecting to
the server by SSH.

We can then use this information to create groups in our inventory,
which map to the base distribution. Since we know the default cloud
user for each distro, we can use a group_vars file for each group to
set ansible_user to the correct username.

Thus:

* Disable legacy group generation in the inventory.
* Set an "os_distro" host variable for each server, based on its
  root volume's image metadata.
* Create a keyed group for each distribution, based on os_distro.
* Select the default user based on the group, via a group_vars entry.
* Select the appropriate role for package updates, based on os_distro.